% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrEvrBeta.R
\name{compEvrBeta}
\alias{compEvrBeta}
\title{Compute Extreme Volatility Risk Betas}
\usage{
compEvrBeta(
  factor,
  asset,
  tau1,
  tau2 = NA,
  samp.freq = 12,
  long.horizon = 3,
  reference = factor,
  garch.order = c(1, 1),
  mean.mod = c(1, 0),
  distr.mod = "norm"
)
}
\arguments{
\item{factor}{time-series of returns of some factor
for which we compute the variance}

\item{asset}{time-series of returns of some asset}

\item{samp.freq}{sampling frequency of the data in terms
of number of observations during a year}

\item{long.horizon}{length of the long horizon cycle in years}

\item{reference}{from which time series should the threshold
value for the asset be computed}

\item{garch.order}{order of the GARCH(p, q) model which should
be estimated on the factor series}

\item{mean.mod}{order of the ARMA(p, q) model which should
be estimated on the factor series}

\item{distr.mod}{distribution model of the error term in GARCH,
for more info, consult the rugarch package}

\item{tau.1}{quantile threshold level for the factor's return}

\item{tau.2}{quantile threshold level for the assets's return}
}
\value{
Returns vector of EVR QS betas for long and short horizon.
}
\description{
This function applies the function \code{getQsBeta} to compute
extreme volatility risk (EVR) quantile spectral (QS) betas
between asset return and negative differences of factor variance
for long and short horizon for tau1 and tau2 quantile thresholds.
If tau2 is not provided, then it is calculated so that the asset's
threshold is tau1-quantile of factor's (reference's) return
}
