% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrEvrBeta.R
\name{compQsBeta}
\alias{compQsBeta}
\title{Compute Quantile Spectral Betas}
\usage{
compQsBeta(factor, asset, tau1, tau2)
}
\arguments{
\item{factor}{time-series of returns of some factor}

\item{asset}{time-series of returns of some asset}

\item{tau.1}{quantile threshold level for the factor's return}

\item{tau.2}{quantile threshold level for the asset's return}
}
\value{
Returns vector of QS betas for frequencies between 0 and pi.
}
\description{
This function is high-level implementation of \code{getBeta} function and
computes general quantile spectral (QS) betas between factor
return and asset return for frequencies in (0,pi)
for tau1 and tau2 quantile thresholds.
}
