% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrEvrBeta.R
\name{compTrBeta}
\alias{compTrBeta}
\title{Compute Tail Risk Betas}
\usage{
compTrBeta(
  factor,
  asset,
  tau1,
  tau2 = NA,
  samp.freq = 12,
  long.horizon = 3,
  reference = factor
)
}
\arguments{
\item{factor}{time-series of returns of some factor}

\item{asset}{time-series of returns of some asset}

\item{samp.freq}{sampling frequency of the data in terms
of number of observations during a year}

\item{long.horizon}{length of the long horizon cycle in years}

\item{reference}{time series from which the threshold
value for the asset should be computed}

\item{tau.1}{quantile threshold level for the factor's return}

\item{tau.2}{quantile threshold level for the asset's return}
}
\value{
Returns vector of TR QS betas for long and short horizon.
}
\description{
This function applies the function \code{getQsBeta} to compute
tail risk (TR) quantile spectral (QS) betas
between asset return and factor return for long and short horizon
for tau1 and tau2 quantile thresholds. If tau2 is not provided,
then it is calculated so that the factor's threshold value is the same
as the asset's threshold value.
}
